requirements/dev-requirements.txt
black==22.3.0
ipdb==0.13.9
ipython>=7.31.1
flake8==4.0.1
isort==5.10.1
slotscheck==0.14.0
twine==3.8.0
mypy==0.961
pip-upgrader==1.4.15
wheel==0.38.1
djangorestframework==3.14.0
environs==9.5.0
pytest==8.3.3
pytest-django==4.9.0
